List resources

List resources

hs_resource(
  page = NULL,
  count = NULL,
  creator = NULL,
  author = NULL,
  group = NULL,
  user = NULL,
  owner = NULL,
  from_date = NULL,
  to_date = NULL,
  subject = NULL,
  published = NULL,
  type = NULL,
  full_text_search = NULL,
  edit_permission = NULL,
  coverage_type = NULL,
  coordinates = NULL,
  include_obselete = NULL
)

Arguments

page
(integer) A page number within the paginated result set.

count
(integer) Number of results to return per page.

creator
(character) The first author (name or email)

author
(character vector) List of authors (name or email)

group
(character) A group name (requires edit_permissions = TRUE)

user
(character) Viewable by user (name or email)

owner
(character) Owned by user (name or email)

from_date
(Date) to get a list of resources created on or after this date

to_date
(Date) to get a list of resources created on or before this date

subject
(character vector) Comma separated list of subjects

published
(logical) filter by published resources?

type
(character vector) list of resources of the specified resource types. \ See details for possible values.

full_text_search
(character) get a list of resources with this text

edit_permission
(logical) filter by edit permissions of user/group/owner?

coverage_type
(character) to get a list of resources that fall within the specified spatial coverage boundary \ Either: "box" or "point"

coordinates
(character vector) list of spatial coordinates in the form of (north, south, east, west)

include_obselete
(logical) Include replaced resources?

Value

A tibble with columns:

resource_title - Title of the resource

resource_type - Resource type

resource_id - ID of the resource

abstract - Resource abstract

authors - Authors of resource

creator - Creator of resource

doi - Resource DOI ID

public - Indicates if resource if public

discoverable - Indicates if resource is discoverable

shareable - Indicates if resource is shareable

immutable - Indicates if resource is immutable

published - Indicates if resource is published

date_created - Date resource was created

date_last_updated - Date resource was last updated

bag_url - URL to resource download, can be passed to download.file

science_metadata_url - URL to science metadata

resource_map_url - URL to resource map

resource_url - URL to resource

content_types - Resource content types, see hs_content_types

Details

Available resource types for type:

GenericResource

RasterResource

RefTimeSeriesResource

TimeSeriesResource

NetcdfResource

ModelProgramResource

ModelInstanceResource

ToolResource

SWATModelInstanceResource

GeographicFeatureResource

ScriptResource

CollectionResource

MODFLOWModelInstanceResource

CompositeResource

Also returned by hs_types.